BREAKING_CHANGE: upgrade ag-grid ^33.2.4 (#530)
BREAKING_CHANGE: upgrade ag-grid ^33.2.4

upgrade build to node 20.x
upgrade libs
upgrade ag-grid-react/community to latest ^33.2.4
Changed onRowDragEnd to have an interface not positional props. It also returns a direction of the move now, not a row index. This is misused as a sequence number in easements on survey and needs fixing. The change in interface should alert whomever does the update.
The selection column is no longer called 'selection', but 'ag-Grid-SelectionColumn'. I cannot control this and it may break your existing tests.
BREAKING_CHANGE: upgrade ag-grid ^33.2.4